backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
motor
bcrypt
pymongo
